Disconnected Value Proxy class from usage by added interface
